% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.r
\name{quartohelp_ragnar_store}
\alias{quartohelp_ragnar_store}
\title{Open the Quarto knowledge store}
\usage{
quartohelp_ragnar_store()
}
\value{
A \code{RagnarStore} object.
}
\description{
Returns a \code{ragnar::RagnarStore} containing the Quarto documentation.
}
\keyword{internal}
