% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.r
\name{update_store}
\alias{update_store}
\title{Updates the Quarto knowledge store}
\usage{
update_store()
}
\value{
\code{NULL} invisibly.
}
\description{
Downloads the latest version of the store from the quartohelp GitHub
repository and builds the search index locally.
}
\details{
The download location can be configured with a few environment variables:
\itemize{
\item \code{QUARTOHELP_STORE_URL}: a custom URL to download the store from. The default is to download the latest store
from the \code{t-kalinowski/quartohelp} repository in the \code{store-v2} release.
\item \code{QUARTOHELP_STORE_RELEASE}: the release tag to download the store from. Defaults to \code{store-v2}.
\item \code{QUARTOHELP_STORE_REPOSITORY}: the repository to download the store from. Defaults to \code{t-kalinowski/quartohelp}.
}
}
